refactor chat widget
