+++ sdk/src/lib/utils.ts
@@ -8,7 +8,12 @@ import type {
   StandardSchemaV1,
   WidgetDefinition,
 } from "./client-types.ts"
-import { INHERITED_PROPERTIES, WIDGET_SLOT_PREFIX, WIDGET_SLOT_SELECTOR } from "./constants.ts"
+import {
+  ASK_QUESTIONNAIRE_TOOL,
+  INHERITED_PROPERTIES,
+  WIDGET_SLOT_PREFIX,
+  WIDGET_SLOT_SELECTOR,
+} from "./constants.ts"
 import type { QuestionnaireItemSpec } from "./types.ts"
 
 export function cn(...inputs: ClassValue[]) {
@@ -94,6 +99,36 @@ export function sanitizeQuestionnaireItems(rawInput: unknown): QuestionnaireItem
 // too; an output-only check would read such a call as still running.
 export function isSettledToolCall(part: { state: string; output?: unknown }): boolean {
   return part.state === "complete" || part.state === "error" || part.output !== undefined
+}
+
+// A busy stream can be silent for a while (server-side reasoning, follow-ups after tool
+// results), so the transcript's "Thinking…" marker shows until the newest assistant part
+// visibly makes progress.
+export function lastPartInProgress(messages: UIMessage[]): boolean {
+  const lastMessage = messages.at(-1)
+  const lastPart = lastMessage?.role === "assistant" ? lastMessage.parts.at(-1) : undefined
+  return lastPart != null && (
+    ((lastPart.type === "text" || lastPart.type === "thinking") &&
+      lastPart.content.length > 0) ||
+    (lastPart.type === "tool-call" && !isSettledToolCall(lastPart))
+  )
+}
+
+// Host tools execute between runs with status "ready". A send in that window ships their
+// call unresolved: the endpoint re-offers the pending tool instead of calling the model,
+// the message goes unanswered, and the redelivered call can re-execute a side-effecting
+// tool — so those windows count as busy too. Questionnaires and calls to tools this mount
+// never implemented stay interactive: the pre-send settle resolves them.
+export function hasPendingToolRun(
+  messages: UIMessage[],
+  toolNames: ReadonlySet<string>,
+): boolean {
+  return messages.some((message) =>
+    message.parts.some((part) =>
+      part.type === "tool-call" && !isSettledToolCall(part) &&
+      part.name !== ASK_QUESTIONNAIRE_TOOL && toolNames.has(part.name)
+    )
+  )
 }
 
 // Transport errors read like "HTTP error! status: 500"; users need something actionable.
